requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.0.5
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-csv==2.1.1
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-pdf==0.2.0
flake8==3.9.2
flake8-broken-line==0.3.0
flake8-isort==4.0.0
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-return==1.1.3
idna==3.2
isort==5.9.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.1.1
pep8-naming==0.12.0
Pillow==8.3.1
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.1.0
python-dotenv==0.18.0
python3-openid==3.2.0
pytz==2021.1
recipes==0.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
testfixtures==6.17.1
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.26.6
